Revert "using float type"
This reverts commit 128d1567734147922fc10b7b7fd6dbeff1354fc9.

+++ src/cicflowmeter/flow.py
@@ -216,7 +216,7 @@ class Flow:
 
         if self.start_timestamp != 0:
             self.flow_interarrival_time.append(
-                float(1e6 * (packet.time - self.latest_timestamp))
+                int(1e6) * (packet.time - self.latest_timestamp)
             )
 
         self.latest_timestamp = max([packet.time, self.latest_timestamp])
